+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - 🆕 overhauled the config system, the mod now creates a directory in the config folder with all configs
 - 🆕 added a tweak that mimics 'Controlling' (to be extended in the future)
 - 🆕 added a tweak to the cases mod to skip the opening animation
-- 🆕 added boss bars to Scape and Run: Parasites (Thanks to thomaslovlin)
+- 🆕 added an offset to the 'scaling health' health renderer
+- 🆕 added boss bars to 'Scape and Run: Parasites' (Thanks to thomaslovlin)
 - 🔄 improved bar renderer with rounded options for all drawables
 - 🔄 improved the description of some config options
 - 🔄 improved instant bone meal tweak to not work on full-grown plants